// next/components/Hero/Hero.d.ts
export interface IPositions {
  c2left: number;
  c2top: number;
  c3left: number;
  c3top: number;
  c4left: number;
  c4top: number;
  c5left: number;
  c5top: number;
  c6left: number;
  c6top: number;
  c7left: number;
  c7top: number;
  c8left: number;
  c8top: number;
  c9left: number;
  c9top: number;
}